+++ lib/features/Orders/presentation/view/widgets/serving_now_item.dart
@@ -32,7 +32,7 @@ class ServingListItem extends StatefulWidget {
   final int Order_item_count;
   final String Client_name;
   final String Client_location;
-  final int Order_Price;
+  final double Order_Price;
   final String Available_time;
 
   @override
